Move extra impls of Metric to their own module
