+++ metriken/Cargo.toml
@@ -15,7 +15,6 @@ repository = "https://github.com/pelikan-io/rustcommon"
 histogram = { version = "0.7.3", path = "../histogram" }
 heatmap = { version = "0.7.5", path = "../heatmap" }
 linkme = "0.3.3"
-once_cell = "1.14.0"
 parking_lot = "0.12.1"
 metriken-derive = { version = "0.1.0", path = "./derive" }
 phf = { version = "0.11.2", features = ["macros"] }
